% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNPS_functions.R
\name{download_GNPS}
\alias{download_GNPS}
\title{download_GNPS}
\usage{
download_GNPS(ID, dir)
}
\arguments{
\item{ID}{The GNPS id to retrieve}

\item{dir}{Location of the output}
}
\value{
Downloads and extracts a file of the GNPS content, returns a list with paths to the buckettable and network files
}
\description{
download_GNPS
}
\examples{
 download_GNPS("0310e20491314ddbbf12d56b592548b4", ".")
}
