% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GEfeatures}
\alias{GEfeatures}
\title{GlobalEuphorbia feature dataset}
\format{Two matrices, one with feature intensities (GEfeatures) and cosine score matrix (GEcss)
\describe{
  \item{GEfeatures)}{Feature intensities}
  \item{GEcss}{Cosine score matrix (generated by GNPS)}
  ...
}}
\source{
\url{https://github.com/DorresteinLaboratory/supplementary-GlobalEuphorbiaStudy}
}
\usage{
GEfeatures
}
\description{
Dataset from Ernst et al. containing tandem MS data from various Euphorbia speciments
}
\keyword{datasets}
