% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNPS_functions.R
\name{prepare_GNPS}
\alias{prepare_GNPS}
\title{prepare_GNPS}
\usage{
prepare_GNPS(ID, dir = ".")
}
\arguments{
\item{ID}{GNPS ID to prepare}

\item{dir}{Path to put GNPS files}
}
\value{
A list of a feature matrix and a chemical similarity matrix (CSS)
}
\description{
prepare_GNPS
}
\examples{
prepare_GNPS("0310e20491314ddbbf12d56b592548b4")
}
