% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cscs.R
\name{distance_cscs}
\alias{distance_cscs}
\title{Calculate CSCS distance}
\usage{
distance_cscs(features, css, dissimilarity = T, cosine_threshold = 0.6,
  weighted = T, verbose = F, normalize = T)
}
\arguments{
\item{features}{Matrix of feature intensities in each sample (pxn)}

\item{css}{Square matrix of cosine similarities of features (pxp)}

\item{dissimilarity}{Output dissimilarity matrix}

\item{cosine_threshold}{Only include features below this threshold}

\item{weighted}{Weight features by intensity (TRUE) or absence/presence (FALSE)}

\item{verbose}{Print extra details}
}
\value{
A pxp matrix of CSCS distances
}
\description{
Calculate CSCS distance
}
\details{
The value of cosine_threshold is 0.6 in Sedio et al. but for certain applications other values might be better.
}
\examples{

#Get arbitrary GNPS project
gnps <- prepare_GNPS("0310e20491314ddbbf12d56b592548b4")
dist <- distance_cscs(gnps$features, gnps$css)

#GlobalEuphorbia data
distance_cscs(GEfeatures, GEcss)
}
\references{
Sedio et al, 2016
}
